pyproject.toml
[tool.poetry]
name = "pipe-syncer"
version = "0.1.0"
description = ""
authors = ["Vladimir Kuznichenkov <kuzaxak.tech@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.31"
PyGithub = "^1.58.1"
semver = "^3.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
parameterized = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
